@@ -33,77 +34,86 @@
     "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@agm/core": "1.0.0-beta.0",
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
-    "@asymmetrik/angular2-leaflet": "2.1.4",
-    "@nebular/auth": "2.0.0-rc.3",
-    "@nebular/theme": "2.0.0-rc.3",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.5",
-    "@swimlane/ngx-charts": "5.2.0",
-    "angular2-chartjs": "^0.3.0",
-    "angular2-toaster": "4.0.0",
-    "bootstrap": "4.0.0-beta.2",
-    "chart.js": "2.5.0",
-    "ckeditor": "4.6.2",
+    "@agm/core": "^1.0.0-beta.5",
+    "@angular/animations": "^7.2.11",
+    "@angular/cdk": "^7.3.6",
+    "@angular/common": "^7.2.11",
+    "@angular/compiler": "^7.2.11",
+    "@angular/core": "^7.2.11",
+    "@angular/forms": "^7.2.11",
+    "@angular/http": "^7.2.11",
+    "@angular/platform-browser": "^7.2.11",
+    "@angular/platform-browser-dynamic": "^7.2.11",
+    "@angular/router": "^7.2.11",
+    "@asymmetrik/ngx-leaflet": "3.0.1",
+    "@nebular/auth": "3.4.2",
+    "@nebular/bootstrap": "3.4.2",
+    "@nebular/security": "3.4.2",
+    "@nebular/theme": "3.4.2",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@swimlane/ngx-charts": "^10.0.0",
+    "angular-tree-component": "7.2.0",
+    "angular2-chartjs": "0.4.1",
+    "angular2-toaster": "^7.0.0",
+    "bootstrap": "4.0.0",
+    "chart.js": "2.7.1",
+    "ckeditor": "4.7.3",
     "classlist.js": "1.1.20150312",
-    "core-js": "^2.4.1",
-    "d3": "4.8.0",
-    "font-awesome": "4.7.0",
+    "core-js": "2.5.1",
+    "echarts": "^4.0.2",
+    "eva-icons": "^1.1.0",
     "intl": "1.2.5",
     "ionicons": "2.0.1",
-    "leaflet": "1.0.3",
-    "nebular-icons": "1.0.6",
-    "ng2-ckeditor": "^1.1.9",
-    "ng2-smart-table": "1.1.0",
-    "ng2-tree": "^2.0.0-alpha.10",
-    "ngx-echarts": "1.2.2",
+    "leaflet": "1.2.0",
+    "nebular-icons": "1.1.0",
+    "ng2-ckeditor": "^1.2.2",
+    "ng2-completer": "2.0.8",
+    "ng2-smart-table": "1.3.5",
+    "ngx-echarts": "^4.0.1",
+    "node-sass": "^4.12.0",
     "normalize.css": "6.0.0",
-    "pace-js": "^1.0.2",
+    "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "^5.4.2",
-    "socicon": "^3.0.5",
-    "tether": "1.4.0",
+    "rxjs": "6.4.0",
+    "rxjs-compat": "6.3.0",
+    "socicon": "3.0.5",
     "tinymce": "4.5.7",
+    "tslib": "^1.9.0",
     "typeface-exo": "0.0.22",
     "web-animations-js": "2.2.5",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.9",
-    "@angular/compiler-cli": "^4.4.6",
-    "@angular/language-service": "^4.4.6",
-    "@compodoc/compodoc": "^1.0.1",
-    "@types/d3-color": "1.0.4",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/leaflet": "1.0.60",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "conventional-changelog-cli": "^1.3.4",
+    "@angular-devkit/build-angular": "0.12.1",
+    "@angular/compiler-cli": "^7.2.11",
+    "@angular/cli": "^7.3.7",
+    "@angular/language-service": "7.2.11",
+    "@compodoc/compodoc": "1.0.1",
+    "@fortawesome/fontawesome-free": "^5.2.0",
+    "@types/d3-color": "1.0.5",
+    "@types/googlemaps": "^3.30.4",
+    "@types/jasmine": "2.5.54",
+    "@types/jasminewd2": "2.0.3",
+    "@types/leaflet": "1.2.3",
+    "@types/node": "6.0.90",
+    "codelyzer": "^4.5.0",
+    "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "jasmine-core": "2.6.4",
+    "jasmine-spec-reporter": "4.1.1",
+    "karma": "1.7.1",
+    "karma-chrome-launcher": "2.1.1",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "1.3.0",
+    "karma-jasmine": "1.1.0",
+    "karma-jasmine-html-reporter": "0.2.2",
     "npm-run-all": "4.0.2",
-    "protractor": "~5.1.2",
+    "protractor": "5.1.2",
     "rimraf": "2.6.1",
     "stylelint": "7.13.0",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "tslint-language-service": "0.9.6",
-    "typescript": "~2.3.3"
+    "ts-node": "3.2.2",
+    "tslint": "^5.17.0",
+    "tslint-language-service": "^0.9.9",
+    "typescript": "3.2.4"
   }
 }
