@@ -1,4 +1,6 @@
-import { Component } from '@angular/core';
+import { Component, OnDestroy } from '@angular/core';
+import { NbThemeService } from '@nebular/theme';
+import { takeWhile } from 'rxjs/operators';
 
 // TODO: move layouts into the framework
 @Component({
@@ -11,9 +13,9 @@ import { Component } from '@angular/core';
       </nb-layout-header>
 
       <nb-sidebar class="menu-sidebar" tag="menu-sidebar" responsive >
-        <nb-sidebar-header>
-          <a href="https://github.com/Kmilo9433/arka4" target="_blank" class="btn btn-hero-success main-btn">
-            <i class="ion ion-social-github"></i> <span>Repositorio</span>
+        <nb-sidebar-header *ngIf="currentTheme !== 'corporate'">
+          <a href="#" class="btn btn-hero-success main-btn">
+            <i class="ion ion-social-github"></i> <span>Support Us</span>
           </a>
         </nb-sidebar-header>
         <ng-content select="nb-menu"></ng-content>
@@ -22,7 +24,7 @@ import { Component } from '@angular/core';
       <nb-layout-column class="small">
       </nb-layout-column>
 
-      <nb-layout-column right>
+      <nb-layout-column>
         <ng-content select="router-outlet"></ng-content>
       </nb-layout-column>
 
@@ -33,5 +35,21 @@ import { Component } from '@angular/core';
     </nb-layout>
   `,
 })
-export class TwoColumnsLayoutComponent {
+export class TwoColumnsLayoutComponent implements OnDestroy {
+
+  private alive = true;
+
+  currentTheme: string;
+
+  constructor(protected themeService: NbThemeService) {
+    this.themeService.getJsTheme()
+      .pipe(takeWhile(() => this.alive))
+      .subscribe(theme => {
+        this.currentTheme = theme.name;
+    });
+  }
+
+  ngOnDestroy() {
+    this.alive = false;
+  }
 }
