@@ -1,50 +1,48 @@
-import { AfterViewInit, Component, OnDestroy } from '@angular/core';
+import { delay, takeWhile } from 'rxjs/operators';
+import { AfterViewInit, Component, Input, OnDestroy } from '@angular/core';
 import { NbThemeService } from '@nebular/theme';
-
-declare const echarts: any;
+import { LayoutService } from '../../../../@core/utils';
+import { ElectricityChart } from '../../../../@core/data/electricity';
 
 @Component({
   selector: 'ngx-electricity-chart',
   styleUrls: ['./electricity-chart.component.scss'],
   template: `
-    <div echarts [options]="option" class="echart"></div>
+    <div echarts
+         [options]="option"
+         class="echart"
+         (chartInit)="onChartInit($event)">
+    </div>
   `,
 })
 export class ElectricityChartComponent implements AfterViewInit, OnDestroy {
 
+  private alive = true;
+
+  @Input() data: ElectricityChart[];
+
   option: any;
-  data: Array<any>;
-  themeSubscription: any;
+  echartsIntance: any;
 
-  constructor(private theme: NbThemeService) {
-
-    const points = [490, 490, 495, 500, 505, 510, 520, 530, 550, 580, 630,
-      720, 800, 840, 860, 870, 870, 860, 840, 800, 720, 200, 145, 130, 130,
-      145, 200, 570, 635, 660, 670, 670, 660, 630, 580, 460, 380, 350, 340,
-      340, 340, 340, 340, 340, 340, 340, 340];
-
-    // const points = [];
-    // let pointsCount = 100;
-    // let min = -3;
-    // let max = 3;
-    // let xStep = (max - min) / pointsCount;
-    //
-    // for(let x = -3; x <= 3; x += xStep) {
-    //   let res = x**3 - 5*x + 17;
-    //   points.push(Math.round(res * 25));
-    // }
-
-    this.data = points.map((p, index) => ({
-      label: (index % 5 === 3) ? `${Math.round(index / 5)}` : '',
-      value: p,
-    }));
+  constructor(private theme: NbThemeService,
+              private layoutService: LayoutService) {
+    this.layoutService.onChangeLayoutSize()
+      .pipe(
+        takeWhile(() => this.alive),
+      )
+      .subscribe(() => this.resizeChart());
   }
 
   ngAfterViewInit(): void {
-    this.themeSubscription = this.theme.getJsTheme().delay(1).subscribe(config => {
-      const eTheme: any = config.variables.electricity;
+    this.theme.getJsTheme()
+      .pipe(
+        takeWhile(() => this.alive),
+        delay(1),
+      )
+      .subscribe(config => {
+        const eTheme: any = config.variables.electricity;
 
-      this.option = {
+        this.option = {
           grid: {
             left: 0,
             top: 0,
@@ -81,10 +79,8 @@ export class ElectricityChartComponent implements AfterViewInit, OnDestroy {
               show: false,
             },
             axisLabel: {
-              textStyle: {
-                color: eTheme.xAxisTextColor,
-                fontSize: 18,
-              },
+              color: eTheme.xAxisTextColor,
+              fontSize: 18,
             },
             axisLine: {
               lineStyle: {
@@ -185,7 +181,17 @@ export class ElectricityChartComponent implements AfterViewInit, OnDestroy {
     });
   }
 
+  onChartInit(echarts) {
+    this.echartsIntance = echarts;
+  }
+
+  resizeChart() {
+    if (this.echartsIntance) {
+      this.echartsIntance.resize();
+    }
+  }
+
   ngOnDestroy() {
-    this.themeSubscription.unsubscribe();
+    this.alive = false;
   }
 }
