@@ -1,38 +1,35 @@
-import { Component } from '@angular/core';
+import { Component, OnDestroy } from '@angular/core';
+import { Camera, SecurityCamerasData } from '../../../@core/data/security-cameras';
+import { takeWhile } from 'rxjs/operators';
 
 @Component({
   selector: 'ngx-security-cameras',
   styleUrls: ['./security-cameras.component.scss'],
   templateUrl: './security-cameras.component.html',
 })
-export class SecurityCamerasComponent {
+export class SecurityCamerasComponent implements OnDestroy {
 
-  cameras: any[] = [{
-    title: 'Camera #1',
-    source: 'assets/images/camera1.jpg',
-  }, {
-    title: 'Camera #2',
-    source: 'assets/images/camera2.jpg',
-  }, {
-    title: 'Camera #3',
-    source: 'assets/images/camera3.jpg',
-  }, {
-    title: 'Camera #4',
-    source: 'assets/images/camera4.jpg',
-  }];
+  private alive = true;
 
-  selectedCamera: any = this.cameras[0];
-
-  userMenu = [{
-    title: 'Profile',
-  }, {
-    title: 'Log out',
-  }];
-
+  cameras: Camera[];
+  selectedCamera: Camera;
   isSingleView = false;
+
+  constructor(private securityCamerasService: SecurityCamerasData) {
+    this.securityCamerasService.getCamerasData()
+      .pipe(takeWhile(() => this.alive))
+      .subscribe((cameras: Camera[]) => {
+        this.cameras = cameras;
+        this.selectedCamera = this.cameras[0];
+      });
+  }
 
   selectCamera(camera: any) {
     this.selectedCamera = camera;
     this.isSingleView = true;
+  }
+
+  ngOnDestroy() {
+    this.alive = false;
   }
 }
