@@ -1,5 +1,8 @@
 import { Component, OnDestroy } from '@angular/core';
 import { NbThemeService } from '@nebular/theme';
+import { Temperature, TemperatureHumidityData } from '../../../@core/data/temperature-humidity';
+import { takeWhile } from 'rxjs/operators';
+import { forkJoin } from 'rxjs';
 
 @Component({
   selector: 'ngx-temperature',
@@ -8,24 +11,43 @@ import { NbThemeService } from '@nebular/theme';
 })
 export class TemperatureComponent implements OnDestroy {
 
-  temperature = 24;
+  private alive = true;
+
+  temperatureData: Temperature;
+  temperature: number;
   temperatureOff = false;
   temperatureMode = 'cool';
 
-  humidity = 87;
+  humidityData: Temperature;
+  humidity: number;
   humidityOff = false;
   humidityMode = 'heat';
 
   colors: any;
   themeSubscription: any;
 
-  constructor(private theme: NbThemeService) {
-    this.themeSubscription = this.theme.getJsTheme().subscribe(config => {
+  constructor(private theme: NbThemeService,
+              private temperatureHumidityService: TemperatureHumidityData) {
+    this.theme.getJsTheme()
+      .pipe(takeWhile(() => this.alive))
+      .subscribe(config => {
       this.colors = config.variables;
     });
+
+    forkJoin(
+      this.temperatureHumidityService.getTemperatureData(),
+      this.temperatureHumidityService.getHumidityData(),
+    )
+      .subscribe(([temperatureData, humidityData]: [Temperature, Temperature]) => {
+        this.temperatureData = temperatureData;
+        this.temperature = this.temperatureData.value;
+
+        this.humidityData = humidityData;
+        this.humidity = this.humidityData.value;
+      });
   }
 
   ngOnDestroy() {
-    this.themeSubscription.unsubscribe();
+    this.alive = false;
   }
 }
