@@ -4,17 +4,34 @@ import { Routes, RouterModule } from '@angular/router';
 import { FormsComponent } from './forms.component';
 import { FormInputsComponent } from './form-inputs/form-inputs.component';
 import { FormLayoutsComponent } from './form-layouts/form-layouts.component';
+import { DatepickerComponent } from './datepicker/datepicker.component';
+import { ButtonsComponent } from './buttons/buttons.component';
 
 const routes: Routes = [{
   path: '',
   component: FormsComponent,
-  children: [{
-    path: 'inputs',
-    component: FormInputsComponent,
-  }, {
-    path: 'layouts',
-    component: FormLayoutsComponent,
-  }],
+  children: [
+    {
+      path: 'inputs',
+      component: FormInputsComponent,
+    },
+    {
+      path: 'layouts',
+      component: FormLayoutsComponent,
+    },
+    {
+      path: 'layouts',
+      component: FormLayoutsComponent,
+    },
+    {
+      path: 'buttons',
+      component: ButtonsComponent,
+    },
+    {
+      path: 'datepicker',
+      component: DatepickerComponent,
+    },
+  ],
 }];
 
 @NgModule({
@@ -33,4 +50,5 @@ export const routedComponents = [
   FormsComponent,
   FormInputsComponent,
   FormLayoutsComponent,
+  DatepickerComponent,
 ];
