@@ -2,216 +2,62 @@ import { NbMenuItem } from '@nebular/theme';
 
 export const MENU_ITEMS: NbMenuItem[] = [
   {
-    title: 'Dashboard',
-    icon: 'nb-home',
-    link: '/pages/dashboard',
-    home: true,
-  },
-  /*{
-    title: 'FEATURES',
+    title: 'ARKA',
     group: true,
   },
   {
-    title: 'UI Features',
-    icon: 'nb-keypad',
-    link: '/pages/ui-features',
-    children: [
-      {
-        title: 'Buttons',
-        link: '/pages/ui-features/buttons',
-      },
-      {
-        title: 'Grid',
-        link: '/pages/ui-features/grid',
-      },
-      {
-        title: 'Icons',
-        link: '/pages/ui-features/icons',
-      },
-      {
-        title: 'Modals',
-        link: '/pages/ui-features/modals',
-      },
-      {
-        title: 'Typography',
-        link: '/pages/ui-features/typography',
-      },
-      {
-        title: 'Animated Searches',
-        link: '/pages/ui-features/search-fields',
-      },
-      {
-        title: 'Tabs',
-        link: '/pages/ui-features/tabs',
-      },
-    ],
+    title: 'MENÚ',
+    icon: 'nb-star',
   },
   {
-    title: 'Forms',
+    title: 'Acta de Recibido',
     icon: 'nb-compose',
     children: [
       {
-        title: 'Form Inputs',
-        link: '/pages/forms/inputs',
+        title: 'Consulta de Actas de Recibido',
+        link: '/pages/acta_recibido/consulta_acta_recibido',
       },
       {
-        title: 'Form Layouts',
-        link: '/pages/forms/layouts',
+        title: 'Registro de Acta de Recibido',
+        link: '/pages/acta_recibido/registro_acta_recibido',
+      },
+      {
+        title: 'Verificación de Acta de Recibido',
+        link: '/pages/acta_recibido/verificacion_acta_recibido',
       },
     ],
   },
   {
-    title: 'Components',
-    icon: 'nb-gear',
+    title: 'Movimientos',
+    icon: 'nb-shuffle',
     children: [
       {
-        title: 'Tree',
-        link: '/pages/components/tree',
-      }, {
-        title: 'Notifications',
-        link: '/pages/components/notifications',
+        title: 'Solicitud de Baja de Bien',
+        link: '/pages/movimientos/solicitud_baja_bien',
+      },
+      {
+        title: 'Aprobación de Baja de Bien',
+        link: '/pages/movimientos/aprobacion_baja_bien',
       },
     ],
   },
   {
-    title: 'Maps',
-    icon: 'nb-location',
-    children: [
-      {
-        title: 'Google Maps',
-        link: '/pages/maps/gmaps',
-      },
-      {
-        title: 'Leaflet Maps',
-        link: '/pages/maps/leaflet',
-      },
-      {
-        title: 'Bubble Maps',
-        link: '/pages/maps/bubble',
-      },
-    ],
-  },
-  {
-    title: 'Charts',
-    icon: 'nb-bar-chart',
-    children: [
-      {
-        title: 'Echarts',
-        link: '/pages/charts/echarts',
-      },
-      {
-        title: 'Charts.js',
-        link: '/pages/charts/chartjs',
-      },
-      {
-        title: 'D3',
-        link: '/pages/charts/d3',
-      },
-    ],
-  },
-  {
-    title: 'Editors',
-    icon: 'nb-title',
-    children: [
-      {
-        title: 'TinyMCE',
-        link: '/pages/editors/tinymce',
-      },
-      {
-        title: 'CKEditor',
-        link: '/pages/editors/ckeditor',
-      },
-    ],
-  },
-  {
-    title: 'Tables',
-    icon: 'nb-tables',
-    children: [
-      {
-        title: 'Smart Table',
-        link: '/pages/tables/smart-table',
-      },
-    ],
-  },
-  {
-    title: 'Auth',
-    icon: 'nb-locked',
-    children: [
-      {
-        title: 'Login',
-        link: '/auth/login',
-      },
-      {
-        title: 'Register',
-        link: '/auth/register',
-      },
-      {
-        title: 'Request Password',
-        link: '/auth/request-password',
-      },
-      {
-        title: 'Reset Password',
-        link: '/auth/reset-password',
-      },
-    ],
-  },
-  */{
-    title: 'MÓDULO CATÁLOGOS',
-    group: true,
-  },
-  {
-    title: 'Grupos',
+    title: 'Entradas',
     icon: 'nb-compose',
     children: [
       {
-        title: 'Crear Grupos',
-        link: '/pages/grupos/crear',
-      },
-      {
-        title: 'Gestionar Grupos',
-        link: '/pages/grupos/gestionar',
+        title: 'Registrar',
+        link: '/pages/entradas/registro',
       },
     ],
   },
   {
-    title: 'Subgrupos',
-    icon: 'nb-compose',
+    title: 'Salidas',
+    icon: 'nb-paper-plane',
     children: [
       {
-        title: 'Crear Subgrupos',
-        link: '/pages/subgrupos/crear',
-      },
-      {
-        title: 'Gestionar Subgrupos',
-        link: '/pages/subgrupos/gestionar',
-      },
-    ],
-  },
-  {
-    title: 'Catalogo',
-    icon: 'nb-compose',
-    children: [
-      {
-        title: 'Gestionar Catalogo',
-        link: '/pages/catalogo/gestionar',
-      },
-    ],
-  },
-  {
-    title: 'MÓDULO ACTA INICIO',
-    group: true,
-  },
-  {
-    title: 'Acta Inicio',
-    icon: 'nb-compose',
-    children: [
-      {
-        title: 'Registrar Acta Inicio',
-        link: '/pages/acta/crear',
-      },
-      {
-        title: 'Gestionar Acta Inicio',
-        link: '/pages/acta/gestionar',
+        title: 'Registrar Salida',
+        link: '/pages/salidas/registro',
       },
     ],
   },
