@@ -1,44 +1,41 @@
 import { RouterModule, Routes } from '@angular/router';
 import { NgModule } from '@angular/core';
-
-import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
+import { PagesComponent } from './pages.component';
+import { NotFoundComponent } from './miscellaneous/not-found/not-found.component';
 
 const routes: Routes = [{
   path: '',
   component: PagesComponent,
-  children: [{
-    path: 'dashboard',
-    component: DashboardComponent,
-  }, {
-    path: 'ui-features',
-    loadChildren: './ui-features/ui-features.module#UiFeaturesModule',
-  }, {
-    path: 'components',
-    loadChildren: './components/components.module#ComponentsModule',
-  }, {
-    path: 'maps',
-    loadChildren: './maps/maps.module#MapsModule',
-  }, {
-    path: 'charts',
-    loadChildren: './charts/charts.module#ChartsModule',
-  }, {
-    path: 'editors',
-    loadChildren: './editors/editors.module#EditorsModule',
-  }, {
-    path: 'forms',
-    loadChildren: './forms/forms.module#FormsModule',
-  }, {
-    path: 'grupos',
-    loadChildren: './grupos/grupos.module#GruposModule',
-  }, {
-    path: 'tables',
-    loadChildren: './tables/tables.module#TablesModule',
-  }, {
-    path: '',
-    redirectTo: 'dashboard',
-    pathMatch: 'full',
-  }],
+  children: [
+    {
+      path: 'dashboard',
+      component: DashboardComponent,
+    },
+    {
+      path: 'acta_recibido',
+      loadChildren: './acta-recibido/acta-recibido.module#ActaRecibidoModule',
+    },
+    {
+      path: 'movimientos',
+      loadChildren: './movimientos/movimientos.module#MovimientosModule',
+    },
+    {
+      path: 'entradas',
+      loadChildren: './entradas/entradas.module#EntradasModule',
+    },
+    {
+      path: 'salidas',
+      loadChildren: './salidas/salidas.module#SalidasModule',
+    },
+    {
+      path: '',
+      redirectTo: 'dashboard',
+      pathMatch: 'full',
+    }, {
+      path: '**',
+      component: NotFoundComponent,
+    }],
 }];
 
 @NgModule({
