@@ -2,49 +2,26 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { UiFeaturesComponent } from './ui-features.component';
-import { ButtonsComponent } from './buttons/buttons.component';
 import { GridComponent } from './grid/grid.component';
 import { IconsComponent } from './icons/icons.component';
-import { ModalsComponent } from './modals/modals.component';
 import { TypographyComponent } from './typography/typography.component';
-import { TabsComponent, Tab1Component, Tab2Component } from './tabs/tabs.component';
 import { SearchComponent } from './search-fields/search-fields.component';
 
 const routes: Routes = [{
   path: '',
   component: UiFeaturesComponent,
-  children: [{
-    path: 'buttons',
-    component: ButtonsComponent,
-  }, {
+  children: [ {
     path: 'grid',
     component: GridComponent,
   }, {
     path: 'icons',
     component: IconsComponent,
   }, {
-    path: 'modals',
-    component: ModalsComponent,
-  }, {
     path: 'typography',
     component: TypographyComponent,
   }, {
     path: 'search-fields',
     component: SearchComponent,
-  }, {
-    path: 'tabs',
-    component: TabsComponent,
-    children: [{
-      path: '',
-      redirectTo: 'tab1',
-      pathMatch: 'full',
-    }, {
-      path: 'tab1',
-      component: Tab1Component,
-    }, {
-      path: 'tab2',
-      component: Tab2Component,
-    }],
   }],
 }];
 
